Migrate spectator_client module from spinnaker/google/stackdriver_monitoring
I've refactored it a little during the migration so that it
knows about the endpoints to scrape from a "conf_dir" rather than
having built-in names. This will allow easy migration to HA deployments
where there are different name variations and also allows the microservice
debian packages to export their own conf file when installed as a simple
way for the monitoring daemon to know they are present without having to
manually configure.