Added initial install.
Note the datadog/prometheus/stackdriver directories under install
are independent of the install scripts that invoke them. They are
globbed into this PR for review convienence.

The bash scripts are temporary. This will be replaced with a debian package
and gradle. For now this is carrying forward my prototype as a starting point.