Gradleify
I copied the gradle boilerplate from other spinnaker projects.
I left the travis stuff out. I'd like to include it but not sure
how it works or how to install the python dependencies I need.
Saving that for another PR.

At least from this one, you can ./gradlew test and it runs the tests.
The real value of gradle will be to produce the debian packages later.