Consolidated configuration to one file plus one per source.
This is a little weird because I'm supporting both CLI and config files
for convienence of use. There is a name collision between using simple
names like "prometheus" to enable monitoring on prometheus while wanting
to use "prometheus" as a dictionary in a common config file to scope
the options it introduces there.

To work around that I alias the command line flags to store with a different
key then map those values into the nested dictionary. The other options that
can appear in different places do explicit checks. There arent many.

I think this will make it easier to configure an installed server
since you wont need command-line configuration.