feat(prometheus): Can configure to discover VMs
This install also adds a feature to specify the client host to
override localhost in the config (so it can be polled from outside).
The host was exposed before, only required manual editing.

This also includes some refactoring and cleanup.
Only GCE VM discovery is implemented in this PR.
Adding the other providers is straightforward where supported by prometheus
(which includes Azure, EC2, and Kubernetes)