Refactor how DatadogMetricsService receives arguments, add support for  … passing tags to Datadog externally, add tests to make sure we can maintain parity with old behaviour + have coverage going forward.
This commit implements a new class DatadogArgumentsGenerator, which
correctly comes up with the arguments needed for the
DatadogMetricsService from the numerous options provided to the factory
service. This allows relative decoupling from external sources and how
the DatadogMetricsService internally works.

This commit includes PR feedback and fixes some minor errors.