chore(daemon): Configure whether to use service tags or decorated metric names
Currently metrics are mutated to decorate them with the service name.
Each service does this independentally. The metrics are also "normalized"
which consideres the "statistic" tag and decorates the name and removes it
to accomodate the semantic implications of this tag and its value.

Here we make the service name decoration a configuration option as well
as the ability to inject a service tag so that we can use common metrics
distinguished by the service. This policy is per-provider but not per-metric.
To make it per-provider the global normalizing function is put into a
class and instiated per metric provider where the provider can change
configuration. There is not currently any provider-specific configuration
but that could be added in the future if needed.