fix(daemon): Added options to normalize HA service names to fix prometheus
This requires setting 'use_base_service_name_only' to True.
If so, it will strip the hyphenated decorator from the service name
(e.g. clouddriver-ro becomes clouddriver) when reporting metrics.
This avoids a prometheus bug where '-' is not permitted in the names,
and also is compatible with the standard dashboards.