chore(daemon): Transformations strip account and application tags when per_*
There is a bunch of refactoring here that reimplements the transformation
processor so that it encapsulates the specification dictionary into a
TransformationRule. It adds the old private keys as attributes to the rule
and moves the manipulation logic from the TransformationProcessor into the rule.
It is mostly not that interesting except there are no longer explicit None rule
entries in the lookup table and instead are empty rules.

The "interesting" part here is that account and application tags are stripped
if per_account or per_application is specified. This was actually stripped before
too if the tags were coming from spectator, but not if they were injected in
a transform_tag rule. Now we do that. The refactoring wasnt needed to add that,
but I thought I was going to do more and the refactoring is better for future
changes so left it.

The problem here is that there is no implementation that actually uses the
per_account or per_application, nor is there a way to turn it on or off.
If those attributes are needed, the per_* can be commented out of the specification
and tags explicitly added where desired (i.e. essentially dont use per_ feature
yet). It is here so I can write the meter specifications as intended long term.