chore(experiment): Proposal for changing/standardizing monitoring data model
See spinnaker-monitoring-daemon/experimental/README.md for more info about the proposal.
This PR is a working specification along with new dashboards for Prometheus and Stackdriver.
They are relegated to an "experimental" tag to sit outside the production codepaths,
but are fully working and compatible with production microservice (and daemon) deployments.

There are no Spinnaker code changes involved, only configuration.
The persistent metric data collected will be incompatible with previous production microservices
when choosing to opt-in to using the provided configuration.