chore(stackdriver): Refactored stackdriver descriptor management.
This moves code around so that Stackdriver Metric Descriptor
management is in a new module stackdriver_descriptor and class
MetricDescriptorManager. The old code was in the handler to upsert
descriptors and some additional runtime support in the
stackdriver_service module. The BatchProcessor is also moved there.

This also refactors the upsert descriptor implementation itself so
that it is split into an audit of deviations and remedies acting on
that audit (e.g. create, update desctiptors). This adds delete
old ones to the remedies.

Also added is a new argument "manage_stackdriver_descriptors" which
controls what the remediation is allowed to do. By default "none"
meaning only report but not act. The upsert command defaults to "create"
consistent with the old behavior (create and update but not delete).

Adds a janitor to Stackdriver that can manage the stackdriver descriptors
(create/update and/or delete depending on mode) periodically until they
appear to be stable. This is to facilitate migrating to the specified
descriptors without needing special installation procedures.