refactor(config): remove unused spinnaker-monitoring-third-party halconfig
The spinnaker-monitoring-third-party package does not actually have any Halyard-generated config, and in fact I can't find any references to spinnaker-monitoring-third-party.yml anywhere in this repository.
Based on documentation, spinnaker-monitoring-third-party needs to be installed as a debian package, and then users are expected to run one or more of the dashboard installation scripts contained in the package to install third-party dashboards.